Joint_Movement_Controller // katana_teleop
i created a katana_teleop package, which shall provide a teleoperation node for the katana
via keyboard (ps3, joystick might follow). the node works in principle but needs to be adjusted to be stable and user friendly.

additonally i implemented a JointMovementActionController capable of adressing
single joints of the katana via a desired JointState as goal.

the implementation of the controller is still in progress.